pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Stephen Sadowski", email = "stephen.sadowski@sjsadowski.com"},
]
dependencies = [
    "sanic[ext]>=22.3.1",
    "python-dotenv>=0.20.0",
    "firebase-admin>=5.2.0",
    "google-cloud-firestore>=2.4.0",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
